IRIS-4942 | move away from using PageWithPosts interface in favour of abstract class
