+++ src/test/java/com/wikia/webdriver/elements/mercury/pages/discussions/PageWithPosts.java
@@ -1,13 +1,28 @@
 package com.wikia.webdriver.elements.mercury.pages.discussions;
 
+
 import com.wikia.webdriver.elements.mercury.components.discussions.common.Post;
+import com.wikia.webdriver.elements.mercury.components.discussions.common.PostEntity;
 import com.wikia.webdriver.elements.mercury.components.discussions.common.SignInToFollowModalDialog;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.WikiBasePageObject;
 
+import java.util.Optional;
 
-public interface PageWithPosts {
+public abstract class PageWithPosts extends WikiBasePageObject {
 
-  Post getPost();
+  public Optional<PostEntity> getPostById(String postId) {
+    return Optional.ofNullable(
+      getPost()
+        .getPosts()
+        .stream()
+        .peek(postEntity -> scrollTo(postEntity.getWebElement()))
+        .filter(postEntity -> postEntity.findId().equals(postId))
+        .findFirst()
+        .orElseThrow(() ->
+          new RuntimeException(String.format("Post with id [%s] not found on page", postId))));
+  }
 
-  SignInToFollowModalDialog getSignInToFollowModalDialog();
+  public abstract Post getPost();
 
+  public abstract SignInToFollowModalDialog getSignInToFollowModalDialog();
 }
