+++ src/test/java/com/wikia/webdriver/elements/mercury/pages/discussions/PostDetailsPage.java
@@ -7,7 +7,7 @@ import com.wikia.webdriver.pageobjectsfactory.pageobject.WikiBasePageObject;
 import lombok.Getter;
 
 
-public class PostDetailsPage extends WikiBasePageObject implements PageWithPosts {
+public class PostDetailsPage extends PageWithPosts {
 
   @Getter(lazy = true)
   private final TopNoteModalDialog topNoteModalDialog = new TopNoteModalDialog();
